backend/requirements.txt
python-dotenv
typer
fastapi
fastapi[standard]
pydantic
elevenlabs
srt
openai-whisper
beautifulsoup4
markdownify
soundfile
openai
pandas
lmnt
arxiv
ffmpeg
joblib
instructor
google-generativeai
groq
deepgram-sdk
mlx_whisper
gradio
